% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_today.R
\name{BD_today}
\alias{BD_today}
\title{Today's date in two formats}
\usage{
BD_today(date = Sys.Date())
}
\arguments{
\item{date}{A date (or something coercible to \code{Date}). Defaults to \code{Sys.Date()}.}
}
\value{
A named character vector of length 2:
\describe{
\item{YMD_underscores}{e.g., "2025_09_16"}
\item{YMD_compact}{e.g., "20250916"}
}
}
\description{
Returns today's date as "YYYY_MM_DD" and "YYYYMMDD".
}
\examples{
BD_today()
BD_today("2024-01-05")

}
